Add coverage metrics to the project (#108)
* Allow running `manage.py test` to run all tests

This file was not used, and confused manage.py test discovery,
requiring the tests directory to be specified every time. Without
this file, the tests directory and contents are auto-discovered.

* Define the settings module for pytest-django

To run all tests, activate the venv and run `pytest`

* Add coverage reporting to pytests

The coverage package gives us the ability to see what lines of code are
actually run by our tests, and which lines are left out. This is first
and foremost useful to see what tests we might need to write next.

- Added `coverage` package to `requirements.txt`, along with the pytest
  plugin to run it automatically with the tests
- Added options to `pytest.ini` to keep coverage data automatically when
  pytests are run, and to output an HTML report of coverage to htmlcov
- New file: `.coveragerc` tells `coverage` to ignore framework code and
  other files that we don't need to test. Also, for reporting don't list
  all the files that are 100% covered, so we can see what work remains